+++ main/src/app/(auth)/login/page.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from "react";
 import { signIn, useSession } from "next-auth/react";
 import { useRouter } from "next/navigation";
 import { motion } from "framer-motion";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/src/components/ui/button";
 import { UtensilsIcon } from "lucide-react";
 
 export default function LoginPage() {
