+++ main/src/app/chat/components/InputArea.tsx
@@ -1,5 +1,5 @@
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
+import { Button } from "@/src/components/ui/button";
+import { Input } from "@/src/components/ui/input";
 import { SendIcon, Loader2Icon } from "lucide-react";
 
 type InputAreaProps = {
@@ -26,22 +26,6 @@ export default function InputArea({
           className="w-full bg-gray-100 dark:bg-gray-800 border-none text-gray-900 dark:text-white pl-4 pr-20 py-7 rounded-full text-sm"
         />
         <div className="absolute right-2 top-1/2 transform -translate-y-1/2 flex items-center space-x-2">
-          {/* <Button
-            type="button"
-            size="icon"
-            variant="ghost"
-            className="text-gray-500 dark:text-gray-400"
-          >
-            <ImageIcon className="h-5 w-5" />
-          </Button>
-          <Button
-            type="button"
-            size="icon"
-            variant="ghost"
-            className="text-gray-500 dark:text-gray-400"
-          >
-            <MicIcon className="h-5 w-5" />
-          </Button> */}
           <Button
             type="submit"
             size="icon"
